closure

+++ application/tool/controller/Detailstatic.php
@@ -303,7 +303,7 @@ class Detailstatic extends Common
                     $main_site, $partnersite, $commonjscode,
                     $article_list, $question_list, $scatteredarticle_list) = Commontool::getEssentialElement('detail', $item->question, $temp_content, $a_keyword_id);
                 $assign_data = compact('com_name', 'title', 'keyword', 'description', 'm_url', 'redirect_code', 'menu', 'before_head', 'after_head', 'chain_type', 'next_site', 'main_site', 'common_site', 'partnersite', 'commonjscode', 'article_list', 'question_list', 'scatteredarticle_list');
-//                file_put_contents('log/article.txt', $this->separator . date('Y-m-d H:i:s') . print_r($assign_data, true) . $this->separator, FILE_APPEND);
+                file_put_contents('log/article.txt', $this->separator . date('Y-m-d H:i:s') . print_r($assign_data, true) . $this->separator, FILE_APPEND);
                 //页面中还需要填写隐藏的 表单 node_id site_id
                 //获取上一篇和下一篇
 
@@ -317,10 +317,12 @@ class Detailstatic extends Common
                         'next_question' => $next_question
                     ]
                 );
-                file_put_contents("11122.txt","111");
-
+                //判断模板是否存在
+                if (!file_exists('question')) {
+                    $this->make_error("question");
+                    die;
+                }
                 $make_web = file_put_contents('question/question' . $item["id"] . '.html', $content);
-                file_put_contents("222.txt",$make_web);die;
                 //开始同步数据库
                 if ($make_web) {
                     $articleCountModel = ArticleSyncCount::where($where)->find();
@@ -337,9 +339,6 @@ class Detailstatic extends Common
                         $articleCountModel->save();
                     }
                     $limit = $item["id"];
-                } else {
-                    $this->make_error("question");
-                    exit();
                 }
             }
         }
